docs: Generate  server command reference (#4795)
* Generate  server command reference docs

* Insert newline

* undo iinsert newline


* * Renane tab to `Server Configuration Parameters`
* Generate argocd-util command docs


* Tweak erver command  descriptions


* Minor tweaks to argocd-util command descriptions


* Move main_test.go to secrets_redactor_test.go into commands package

* Disable auto generation tag.
