+++ docs/operator-manual/server-commands/argocd-util_apps.md
@@ -1,0 +1,24 @@
+## argocd-util apps
+
+Utility commands operate on ArgoCD applications
+
+### Synopsis
+
+Utility commands operate on ArgoCD applications
+
+```
+argocd-util apps [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for apps
+```
+
+### SEE ALSO
+
+* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
+* [argocd-util apps diff-reconcile-results](argocd-util_apps_diff-reconcile-results.md)	 - Compare results of two reconciliations and print diff.
+* [argocd-util apps get-reconcile-results](argocd-util_apps_get-reconcile-results.md)	 - Reconcile all applications and stores reconciliation summary in the specified file.
+
