+++ docs/operator-manual/server-commands/argocd-util_apps_diff-reconcile-results.md
@@ -1,0 +1,22 @@
+## argocd-util apps diff-reconcile-results
+
+Compare results of two reconciliations and print diff.
+
+### Synopsis
+
+Compare results of two reconciliations and print diff.
+
+```
+argocd-util apps diff-reconcile-results PATH1 PATH2 [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for diff-reconcile-results
+```
+
+### SEE ALSO
+
+* [argocd-util apps](argocd-util_apps.md)	 - Utility commands operate on ArgoCD applications
+
