+++ docs/operator-manual/server-commands/argocd-util_version.md
@@ -1,0 +1,23 @@
+## argocd-util version
+
+Print version information
+
+### Synopsis
+
+Print version information
+
+```
+argocd-util version [flags]
+```
+
+### Options
+
+```
+  -h, --help    help for version
+      --short   print just the version number
+```
+
+### SEE ALSO
+
+* [argocd-util](argocd-util.md)	 - argocd-util tools used by Argo CD
+
